# Consider known predicates in Booster's simplifier API