--- a/booster/test/rpc-integration/test-use-path-condition-in-equations/README.md
+++ b/booster/test/rpc-integration/test-use-path-condition-in-equations/README.md
@@ -1,0 +1,3 @@
+# Test the use of the known path condition when applying simplifications
+
+See `../resourses/use-path-condition-in-equations.k`.