… booster/test/rpc-integration/test-use-path-condition-in-equations/response-test2.json