@geo2a 2024-06-20
KEVM performance:
| Test                                         | georgy-check-constraint-with-known time | master-fb59a8493 time | (georgy-check-constraint-with-known/master-fb59a8493) time
|----------------------------------------------|-----------------------------------------|-----------------------|------------------------------------------------------------
| erc20/ds/balanceOf-spec.k                    | 60.16                                   | 63.91                 | 0.9413237365044594
| erc20/ds/approve-failure-spec.k              | 70.64                                   | 73.41                 | 0.962266721155156
| kontrol/test-storetest-testaccesses-0-spec.k | 84.24                                   | 77.61                 | 1.085427135678392
| TOTAL                                        | 215.04000000000002                      | 214.93                | 1.0005117945377566