@geo2a 2024-06-21
To my disappointment, there still seems to be no difference:

| Test                               | georgy-check-constraint-with-known time | master-fb59a8493 time | (georgy-check-constraint-with-known/master-fb59a8493) time
|------------------------------------|-----------------------------------------|-----------------------|------------------------------------------------------------
| benchmarks/functional-spec.k       | 596.33                                  | 575.15                | 1.036825176041033
| erc20/ds/balanceOf-spec.k          | 63.79                                   | 60.94                 | 1.0467673121102725
| erc20/ds/approve-success-spec.k    | 92.41                                   | 86.48                 | 1.0685707678075855
| erc20/ds/approve-failure-spec.k    | 71.99                                   | 67.09                 | 1.073036220002981
| erc20/ds/transfer-success-2-spec.k | 94.19                                   | 86.76                 | 1.0856385431074227
| TOTAL                              | 918.71                                  | 876.42                | 1.0482531206499168