@jberthold 2024-06-24
> ... a simple integration test that demonstrates the utility...

I was trying something related (but not quite sufficient) recently, using code like this:
```
  configuration
    <k> $PGM:Thing ~> .K </k>
    <i> 42 </i>
  rule <k>AA(X) => BB(X) ... </k>
       <i> I </i>
    requires X ==Int I
  rule AA(X) => CC(X) [owise]
```
I tested that we are able to rewrite `AA(X)` to  `BB(X)` if side conditions _imply_ (but not state syntactically) that `X == 42`.
Maybe we could test your change with an additional function `f` and `f(X) ==Int I` in the side condition, and simplification `f(X) => X requires X ==Int 42`. A bit contrived, though...