@geo2a 2024-07-09
@jberthold I've added two integration [tests](https://github.com/runtimeverification/haskell-backend/blob/1e56d5cb22a6407e12db2987c3a3045ebf0d4e14/booster/test/rpc-integration/resources/use-path-condition-in-equations.k) that make sure we use the path condition when trying simplifications:
- the first test applies a simplification whose side condition is syntactically present in the path condition. I've checked that this test fails on master;
- the second test is a slight modification of the first one which actually needs SMT to infer the simplification's side condition from the path condition. This test aborts on both master and this PR. It will not abort anymore when we start using the solver when applying equations.